software/src/map-genes/requirements.txt
pandas==2.2.3
polars-lts-cpu==1.32.1